+++ src/App.css
@@ -1,0 +1,10 @@
+.wrapper{
+    width: 100%;
+    height: max-content;
+    /* background-color: black; */
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    
+    padding: 10px;
+}
