+++ src/App.jsx
@@ -1,12 +1,16 @@
 import React from 'react'
-// import { RouterProvider } from 'react-router'
-// import { Element } from './routes/router'
+import { RouterProvider } from 'react-router'
+import { Element } from './routes/router'
 import Nextlevel from './components/Nextlevel'
 import Opportunities from './components/Opportunities'
+import WhyZscout from './components/WhyZscout'
+
 const App = () => {
   return (
-    <div>
-      {/* <RouterProvider router={Element}/> */}
+    <div className='wrapper'>
+      <RouterProvider router={Element}/>
+      
+     <WhyZscout/>
      <Opportunities/>
      <Nextlevel/>
     </div>
