+++ src/components/nextLevel.css
@@ -28,7 +28,7 @@
     padding: 20px;
     border-radius: 30px;
     background-color: navy;
-    background-image: url("../assets/Public/cta.jpeg");
+    background-image: url("../assets/images/cta.jpeg");
     background-position:center;
     background-size: cover;
 }
