+++ src/components/opportunities.css
@@ -10,86 +10,6 @@
 }
 
 
-.Opportunities-connection-growth{
-    
-  width: 100%;
-  height: 400px;
-    /* border: 5px solid black; */
-    padding: 10px;
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-    p{
-        font-family: Poppins;
-        font-size: 18px;
-        line-height: 24px;
-         color:rgba(12, 143, 0, 1) ; 
-          padding-left: 9.5%;
-        
-    }
-    h4{
-        font-family: sans-serif;
-        font-weight: 700;
-        font-size: 32px;
-        line-height: 24px;
-        padding-left: 9.5%;
-
-    }
-
-   
-}
-
-.o-c-g-wrapper{
-    width: 100%;
-    height: 340px;
-    /* border: 5px solid red; */
-    background-repeat:no-repeat ;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    gap: 2%;
-}
-.o-c-g{
-    width: 19%;
-    height: 80%;
-    border: 1.7px solid black;
-    border-radius: 20px;
-    display: flex;
-    flex-direction: column;
-    /* gap: 3%; */
-    align-items: center;
-    justify-content: center;
-    
-    h5{
-        /* font-size: 18px; */
-        font-family: Poppins;
-        font-weight: 900;
-        font-size: 17.96px;
-        line-height: 22.44px;
-        letter-spacing: 0px;
-        vertical-align: middle;
-
-    }
-    p{
-        text-align: center;
-        color: rgba(0, 0, 0, 0.766) !important;
-        font-size: 14px;
-        padding: 10px;
-        /* background-color: blue; */
-    }
-}
-
-.o-c-g-IconHolder{
-    width: 19%;
-    height: 20%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    border-radius: 50px;
-    background-color: rgba(159, 255, 150, 1);
-}
-
-
 .HowItWorks{
     width: 100%;
     height: max-content;
@@ -152,62 +72,6 @@
     }
 }
 
-.nextLevel{
-    width: 100%;
-    height: 400px;
-    /* border: 5px solid blue; */
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.nextLevelImage{
-    width: 84%;
-    height: 92%;
-    display: flex;
-    align-items: last baseline;
-    justify-content: space-around;
-    padding: 20px;
-    border-radius: 30px;
-    background-color: navy;
-    background-image: url("../assets/Public/cta.jpeg");
-    background-position:center;
-    background-size: cover;
-}
-
-.nextLevelText1{
-    width: 36%;
-    height: 60%;
-    display: flex;
-    gap: 5%;
-    padding: 10px;
-    flex-direction: column;
-    justify-content: center;
-    /* border: 5px solid white; */
-    h5{
-        color: white;
-        font-size: 35px;
-        word-spacing: 5%;
-    }
-    p{
-        color: white;
-        font-size: 17px;
-    }
-}
-
-.nextLevelText2{
-    width: 20% ;
-    height: 15%;
-    border-radius: 10px;
-    border: none;
-    color: white;
-    background-color: rgb(7,35,47);
-    font-size: 17px;
-}
-
-.nextLevelText2:hover{
-    background-color: rgba(23, 152, 11, 0.879);
-}
 
 
 
@@ -228,56 +92,7 @@
     
   }
 
-  .Opportunities-connection-growth{
-    width: 100%;
-    height:max-content;
-    border: 5px solid black;
-    display: flex;
-    flex-direction: column;
-    p{
-        font-family: Poppins;
-        font-size: 10px;
-         color:rgba(12, 143, 0, 1) ; 
-          padding-left: 9.5%;
-        
-    }
-    h4{
-        font-family: sans-serif;
-        font-size: 10px ;
-        padding-left: 9.5%;
-
-    }
-  }
-
-  .o-c-g-wrapper{
-    width: 100%;
-    height: auto;
-    display: grid;
-    grid-template-columns: 2fr 2fr;
-  }
-
-  .o-c-g{
-    width: 100%;
-    height: 90%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    padding: 5px;
-    padding-top: 20px;
-    h5{
-        font-size: 10px;
-        width: 100%;
-        margin: 0;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-    }
-    p{
-        font-size: 10px !important;
-         text-align: center;
-         
-    }
-  }
+ 
 
 
   .HowItWorks{
@@ -315,35 +130,5 @@ p{
 }
   }
 
-  .nextLevel{
-    width: 100%;
-    height: 250px;
-    display: flex;
-    /* border: 5px solid red; */
-  }
-  .nextLevelImage{
-    width: 100%;
-    height: 90%;
-    display: flex;
-    align-items: center;
-    flex-direction: column;
-  }
-
-  .nextLevelText1{
-    display: flex;
-    flex-direction: column;
-    gap: 2%;
-    width: 92%;
-    h5{
-        font-size: 25px;
-        text-align: center;
-    }
-    p{
-        font-size:12px ;
-    }
-  }
-  .nextLevelText2{
-    width: 80%;
-    font-size: 15px;
-  }
+ 
 }
