+++ src/components/whyZscout.css
@@ -1,0 +1,151 @@
+.whyZscout-wrapper{
+    width: 100%;
+    height: max-content;
+    /* background-color: black; */
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    
+    padding: 10px;
+}
+
+
+
+.Opportunities-connection-growth{
+    
+    width: 100%;
+    height: 400px;
+      /* border: 5px solid black; */
+      padding: 10px;
+      display: flex;
+      flex-direction: column;
+      gap: 10px;
+      p{
+          font-family: Poppins;
+          font-size: 18px;
+          line-height: 24px;
+           color:rgba(12, 143, 0, 1) ; 
+            padding-left: 9.5%;
+          
+      }
+      h4{
+          font-family: sans-serif;
+          font-weight: 700;
+          font-size: 32px;
+          line-height: 24px;
+          padding-left: 9.5%;
+  
+      }
+  
+     
+  }
+  
+  .o-c-g-wrapper{
+      width: 100%;
+      height: 340px;
+      /* border: 5px solid red; */
+      background-repeat:no-repeat ;
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      gap: 2%;
+  }
+  .o-c-g{
+      width: 19%;
+      height: 80%;
+      border: 1.7px solid black;
+      border-radius: 20px;
+      display: flex;
+      flex-direction: column;
+      /* gap: 3%; */
+      align-items: center;
+      justify-content: center;
+      
+      h5{
+          /* font-size: 18px; */
+          font-family: Poppins;
+          font-weight: 900;
+          font-size: 17.96px;
+          line-height: 22.44px;
+          letter-spacing: 0px;
+          vertical-align: middle;
+  
+      }
+      p{
+          text-align: center;
+          color: rgba(0, 0, 0, 0.766) !important;
+          font-size: 14px;
+          padding: 10px;
+          /* background-color: blue; */
+      }
+  }
+  
+  .o-c-g-IconHolder{
+      width: 19%;
+      height: 20%;
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      border-radius: 50px;
+      background-color: rgba(159, 255, 150, 1);
+  }
+
+
+  @media only screen and (max-width: 1440px),
+       only screen and (max-width: 1200px),
+       only screen and (max-width: 900px),
+       only screen and (max-width: 760px),
+       only screen and (max-width: 480px),
+       only screen and (max-width: 379px) {
+        
+  .Opportunities-connection-growth{
+    width: 100%;
+    height:max-content;
+    /* border: 5px solid black; */
+    display: flex;
+    flex-direction: column;
+    p{
+        font-family: Poppins;
+        font-size: 10px;
+         color:rgba(12, 143, 0, 1) ; 
+          padding-left: 9.5%;
+        
+    }
+    h4{
+        font-family: sans-serif;
+        font-size: 10px ;
+        padding-left: 9.5%;
+
+    }
+  }
+
+  .o-c-g-wrapper{
+    width: 100%;
+    height: auto;
+    display: grid;
+    grid-template-columns: 2fr 2fr;
+  }
+
+  .o-c-g{
+    width: 100%;
+    height: 90%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    padding: 5px;
+    padding-top: 20px;
+    h5{
+        font-size: 10px;
+        width: 100%;
+        margin: 0;
+        display: flex;
+        justify-content: center;
+        align-items: center;
+    }
+    p{
+        font-size: 10px !important;
+         text-align: center;
+         
+    }
+  }
+       }
